pyproject.toml
[project]
name = "devscout-automation"
version = "0.1.0"
description = "Automated daily job search for DevScout platform"
dependencies = [
    "playwright>=1.44.0",
    "python-dotenv>=1.0.1",
    "schedule>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]